Fixes #11; use consistent link styling for files to download.
We are rendering links that readers are expected to download as code blocks. The
link here was pointing to a git repository where the expectation on my part may
have been for the student to click the link, but the styling of the rendered link
makes the reader think that they should use `wget`.

Change the link to point directly at the Markdown file.